+++ src/main/java/com/b1n_ry/yigd/events/AllowBlockUnderGraveGenerationEvent.java
@@ -7,12 +7,12 @@ import net.minecraft.block.BlockState;
 
 public interface AllowBlockUnderGraveGenerationEvent {
     Event<AllowBlockUnderGraveGenerationEvent> EVENT = EventFactory.createArrayBacked(AllowBlockUnderGraveGenerationEvent.class, events -> (grave, currentUnder) -> {
+        boolean allow = true;
         for (AllowBlockUnderGraveGenerationEvent event : events) {
-            if (!event.allowBlockGeneration(grave, currentUnder))
-                return false;
+            allow = allow && event.allowBlockGeneration(grave, currentUnder);
         }
 
-        return true;
+        return allow;
     });
 
     boolean allowBlockGeneration(GraveComponent grave, BlockState currentUnder);