+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Through the backup in game GUI you can now also restore items with specific drop rules
 * Updated enchantment configs so that you can configure how the enchantments are accessed.
 Note that if you had soulbound disabled you'll need to disable it again in the config
+* Added grave compass. Configurable in the config file
 
 ### Fixes
 * Game will no longer crash when dying if Travelers backpack is installed